@@ -12,20 +12,21 @@
     },
     "dependencies": {
         "axios": "^1.6.7",
-        "express": "^4.18.2",
+        "cors": "^2.8.5",
+        "express": "^4.18.3",
         "helmet": "^7.1.0",
         "react": "^18.2.0",
         "react-dom": "^18.2.0"
     },
     "devDependencies": {
-        "@types/react": "^18.2.43",
-        "@types/react-dom": "^18.2.17",
+        "@types/react": "^18.2.61",
+        "@types/react-dom": "^18.2.19",
         "@vitejs/plugin-react": "^4.2.1",
-        "eslint": "^8.55.0",
-        "eslint-plugin-react": "^7.33.2",
+        "eslint": "^8.57.0",
+        "eslint-plugin-react": "^7.34.0",
         "eslint-plugin-react-hooks": "^4.6.0",
         "eslint-plugin-react-refresh": "^0.4.5",
         "json-server": "1.0.0-alpha.23",
-        "vite": "^5.0.8"
+        "vite": "^5.1.4"
     }
 }
